Пофиксаны ошибки линтера
