+++ gulpfile.js
@@ -11,19 +11,19 @@ const rename = require(`gulp-rename`);
 const imagemin = require(`gulp-imagemin`);
 const rollup = require(`gulp-better-rollup`);
 const sourcemaps = require(`gulp-sourcemaps`);
-const mocha = require(`gulp-mocha`);                
-const commonjs = require(`rollup-plugin-commonjs`); 
+const mocha = require(`gulp-mocha`);
+const commonjs = require(`rollup-plugin-commonjs`);
 
 gulp.task(`test`, function () {
   return gulp
   .src([`js/**/*.test.js`])
   .pipe(rollup({
     plugins: [
-      commonjs()           // Сообщает Rollup, что модули можно загружать из node_modules
-    ]}, `cjs`))            // Выходной формат тестов — `CommonJS` модуль
+      commonjs()
+    ]}, `cjs`))
   .pipe(gulp.dest(`build/data`))
   .pipe(mocha({
-    reporter: `spec`       // Вид в котором я хочу отображать результаты тестирования
+    reporter: `spec`
   }));
 });
