+++ js/answers/lastGamesResults.js
@@ -1,12 +1,12 @@
 const lastGamesResults = [
-		  { userScore : 4, noteLives : 3, gameTimeLeft : 65 },
-		  { userScore : 5, noteLives : 1, gameTimeLeft : 85 },
-		  { userScore : 8, noteLives : 0, gameTimeLeft : 65 },
-		  { userScore : 11, noteLives : 2, gameTimeLeft : 65 },
-	    { userScore : 4, noteLives : 3, gameTimeLeft : 65 },
-		  { userScore : 5, noteLives : 1, gameTimeLeft : 85 },
-		  { userScore : 8, noteLives : 0, gameTimeLeft : 65 },
-		  { userScore : 11, noteLives : 2, gameTimeLeft : 65 },
-		];
+  {userScore: 4, noteLives: 3, gameTimeLeft: 65},
+  {userScore: 5, noteLives: 1, gameTimeLeft: 85},
+  {userScore: 8, noteLives: 0, gameTimeLeft: 65},
+  {userScore: 11, noteLives: 2, gameTimeLeft: 65},
+  {userScore: 4, noteLives: 3, gameTimeLeft: 65},
+  {userScore: 5, noteLives: 1, gameTimeLeft: 85},
+  {userScore: 8, noteLives: 0, gameTimeLeft: 65},
+  {userScore: 11, noteLives: 2, gameTimeLeft: 65},
+];
 
 export default lastGamesResults;
