+++ js/screens/application.js
@@ -1,7 +1,7 @@
 
-import {getElementFromTemplate, renderScreen, renderWrap, initialGameState, startGame, startCircleTimer, startTimer} from '../utils.js';
+import {renderScreen} from '../utils.js';
 import Model from './model';
-import GameScreen from './gameScreen';
+import GameScreen from './game-screen';
 import {WelcomeScreen} from './welcome';
 import {WinScreen} from './win';
 import SplashScreen from './splash-screen';
@@ -15,14 +15,16 @@ export class Application {
     const splash = new SplashScreen();
     renderScreen(splash.element);
     splash.start();
-    setTimeout(function() {
+    setTimeout(function () {
       window.fetch(`https://es.dump.academy/guess-melody/questions`).
       then((response) => splash.checkStatus(response)).
       then((response) => response.json()).
-      then((data) => gameData = data).
-      then((response) => Application.showWelcome()).
+      then((data) => {
+        gameData = data;
+      }).
+      then(() => Application.showWelcome()).
       catch(Application.showError).
-      then(() => splash.stop())
+      then(() => splash.stop());
     }, 1000);
   }
 
@@ -32,20 +34,20 @@ export class Application {
   }
 
   static showGame() {
-    const model = new Model(gameData); 
+    const model = new Model(gameData);
     const gameScreen = new GameScreen(model);
     gameScreen.startGame();
   }
 
   static showStats(allGamesResults, userResultsScope, userPositionIndex, userComparison, stats, fastAnswers) {
-    const statistics = new WinScreen(allGamesResults, userResultsScope, userPositionIndex, 
-      userComparison, stats, fastAnswers);
+    const statistics = new WinScreen(allGamesResults, userResultsScope, userPositionIndex,
+        userComparison, stats, fastAnswers);
     renderScreen(statistics.element);
   }
 
   static showError(error) {
     const errorScreen = new ErrorScreen(error);
-    renderScreen(errorScreen.element)
+    renderScreen(errorScreen.element);
   }
 
 }
