+++ js/screens/error-screen.js
@@ -1,12 +1,12 @@
 import {AbstractView} from './abstract-view';
 
 export default class ErrorScreen extends AbstractView {
-	constructor(error) {
-		super();
-		this.error = error;
-	}
+  constructor(error) {
+    super();
+    this.error = error;
+  }
 
-	get template() {
+  get template() {
     return `
       <div class="end">
         <p>Произошла ошибка: ${this.error.message}</p>
