+++ js/screens/time.js
@@ -1,16 +1,13 @@
-import {getElementFromTemplate, renderScreen} from '../utils.js';
-import {WelcomeScreen} from './welcome';
 import {AbstractView} from './abstract-view';
 
 export class TimeScreen extends AbstractView {
 
-	render() {
-		return `<div class="main main--result">
-	  	<section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
-	    <h2 class="title">Увы и ах!</h2>
-	    <div class="main-stat">Время вышло!<br>Вы не успели отгадать все мелодии</div>
-	    <span role="button" tabindex="0" class="main-replay">Попробовать ещё раз</span>
-			</div>`
-	};
-
-};
+  render() {
+    return `<div class="main main--result">
+      <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
+      <h2 class="title">Увы и ах!</h2>
+      <div class="main-stat">Время вышло!<br>Вы не успели отгадать все мелодии</div>
+      <span role="button" tabindex="0" class="main-replay">Попробовать ещё раз</span>
+      </div>`;
+  }
+}
