+++ js/screens/win.js
@@ -1,45 +1,39 @@
-import {getElementFromTemplate, renderScreen, userResultsDisplay} from '../utils.js';
-import {WelcomeScreen} from './welcome';
 import {AbstractView} from './abstract-view';
 import {Application} from './application';
 
-
 export class WinScreen extends AbstractView {
-	constructor(allGamesResults, userResultsScope, userPositionIndex, userComparison, state, fastAnswers) {
-		super();
-		this.fastAnswers = fastAnswers;
-		this.state = state; 
-		this.allGamesResults = allGamesResults;
-		this.userResultsScope = userResultsScope;
-		this.userPositionIndex = userPositionIndex;
-		this.userComparison = userComparison;
-	}
+  constructor(allGamesResults, userResultsScope, userPositionIndex, userComparison, state, fastAnswers) {
+    super();
+    this.state = state;
+    this.allGamesResults = allGamesResults;
+    this.userResultsScope = userResultsScope;
+    this.userPositionIndex = userPositionIndex;
+    this.userComparison = userComparison;
+    this.fastAnswers = fastAnswers;
+  }
 
-	render() {
-		return `<div class="main main--result">
-		  <h2 class="title">Вы настоящий меломан!</h2>
-		  <div class="main-stat">За&nbsp;${Math.floor(this.state.timeSpentSec / 60)}&nbsp;минуты и ${this.state.timeSpentSec % 60}&nbsp;секунд
-		    <br>вы&nbsp;набрали ${this.userResultsScope.userScore} баллов (${this.fastAnswers} быстрых)
-		    <br>совершив ${3 - this.userResultsScope.noteLives} ошибки
-		  </div>
-		  <span class="main-comparison">Вы заняли ${this.userPositionIndex + 1} место из ${this.allGamesResults.length}.
-		   Это&nbsp;лучше чем у&nbsp;${this.userComparison()}%&nbsp;игроков</span>
-		  <span role="button" tabindex="0" class="main-replay">Сыграть ещё раз</span>
-			</div>`
-	}
+  render() {
+    return `<div class="main main--result">
+      <h2 class="title">Вы настоящий меломан!</h2>
+      <div class="main-stat">За&nbsp;${Math.floor(this.state.timeSpentSec / 60)}&nbsp;минуты и ${this.state.timeSpentSec % 60}&nbsp;секунд
+        <br>вы&nbsp;набрали ${this.userResultsScope.userScore} баллов (${this.fastAnswers} быстрых)
+        <br>совершив ${3 - this.userResultsScope.noteLives} ошибки
+      </div>
+      <span class="main-comparison">Вы заняли ${this.userPositionIndex + 1} место из ${this.allGamesResults.length}.
+       Это&nbsp;лучше чем у&nbsp;${this.userComparison()}%&nbsp;игроков</span>
+      <span role="button" tabindex="0" class="main-replay">Сыграть ещё раз</span>
+      </div>`;
+  }
 
+  onAnswer() {
+    Application.showWelcome();
+  }
 
-	onAnswer() {
-		gameState.noteLivesMissed = 0;
-		gameState.FIRSTTRACK = 0;
-		application.showWelcome();
-	}
-
-	bind() {
-		this.element.querySelector(`.main-replay`).addEventListener(`click`, () => {
-			this.onAnswer();
-		});
-	}
+  bind() {
+    this.element.querySelector(`.main-replay`).addEventListener(`click`, () => {
+      this.onAnswer();
+    });
+  }
 
 }
